+++ oshi-core/src/main/java/oshi/software/os/unix/freebsd/FreeBsdOperatingSystem.java
@@ -114,7 +114,7 @@ public class FreeBsdOperatingSystem extends AbstractOperatingSystem {
 
     @Override
     public InternetProtocolStats getInternetProtocolStats() {
-        return null;
+        return new FreeBsdInternetProtocolStats();
     }
 
     @Override
